Implement authentication check in AuthLayout and remove redundant checks from LoginPage and SignupPage
